[BWAPI4]  - Remove holiday hack.  - Create BW::Bitmap and BW::Font classes.  - Move several rendering/text functions from dialog/graphics to Bitmap and/or Font.  - Remove several includes of windows.h  - Define NOMINMAX to disable windows macros for min and max (use algorithm/std instead).  - Remove Game::getScreenBuffer.  - Some work on graphics engine interpretation. Terrain, creep, and most images are working (large images don't appear to work correctly?).
git-svn-id: http://bwapi.googlecode.com/svn@4197 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ branches/bwapi4/bwapi/BWAPI/Source/DLLMain.cpp
@@ -24,7 +24,7 @@
 #include "../../Debug.h"
 
 //---------------------------------------------- QUEUE COMMAND -----------------------------------------------
-void __fastcall QueueGameCommand(void *pBuffer, DWORD dwLength)
+void __fastcall QueueGameCommand(void *pBuffer, size_t dwLength)
 {
   if ( !pBuffer || !dwLength || !isCorrectVersion )
     return;
